fix: move interfaces in types file
